requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
python-dotenv==1.0.1
PyJWT==2.10.1
openpyxl==3.1.5
xlrd==2.0.1
Werkzeug==3.1.3
